requirements.txt
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.100.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Pydantic==0.11.0
Flask-RESTful==0.3.10
flask-swagger-ui==4.11.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-model==0.0.5
MarkupSafe==2.1.3
motor==3.2.0
packaging==23.1
pluggy==1.2.0
pydantic==2.0.2
pydantic_core==2.1.2
PyJWT==2.7.0
pymongo==4.4.0
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.7.1
Werkzeug==2.3.6
